+++ lib/typed_eav.rb
@@ -28,6 +28,7 @@ module TypedEAV
   autoload :ValueVersion
   autoload :Versioned
   autoload :Versioning
+  autoload :FieldDeletion
 
   # Raised when a model declared `has_typed_eav scope_method: ...` but no
   # scope can be resolved at query time and `config.require_scope` is truthy.